[WIP] Migration from Ansible playbooks to badfish.
This is the first in a series of patchas to migrate the boot order
logic to be handled with badfish via redfish api.
Ansible playbooks have been removed, tools migrated from bin directory
to python tools.
Added a hard copy of the latest badfish code.
New entrypoints for the api plus additional fields on the mongo db
models.

Change-Id: Ifda6680966aa774ce849815f954f6a804a42f333